+++ addons/filamind_quality_iot/models/__init__.py
@@ -1,0 +1,3 @@
+from . import filamind_quality_point
+from . import filamind_quality_check
+from . import filamind_quality_alert